+++ openbb_platform/extensions/fixedincome/openbb_fixedincome/corporate/__init__.py
@@ -1,0 +1,1 @@
+"""Initialize the Fixed Income Corporate module."""
