+++ openbb_platform/extensions/fixedincome/openbb_fixedincome/government/__init__.py
@@ -1,0 +1,1 @@
+"""Initialize the Fixed Income Government module."""
