+++ openbb_platform/extensions/fixedincome/openbb_fixedincome/government/government_router.py
@@ -1,0 +1,67 @@
+"""Fixed Income Government Router."""
+from openbb_core.app.model.command_context import CommandContext
+from openbb_core.app.model.obbject import OBBject
+from openbb_core.app.provider_interface import (
+    ExtraParams,
+    ProviderChoices,
+    StandardParams,
+)
+from openbb_core.app.query import Query
+from openbb_core.app.router import Router
+from pydantic import BaseModel
+
+router = Router(prefix="/government")
+
+# pylint: disable=unused-argument
+
+
+@router.command(model="USYieldCurve")
+def us_yield_curve(
+    cc: CommandContext,
+    provider_choices: ProviderChoices,
+    standard_params: StandardParams,
+    extra_params: ExtraParams,
+) -> OBBject[BaseModel]:  # type: ignore
+    """US Yield Curve. Get United States yield curve."""
+    return OBBject(results=Query(**locals()).execute())
+
+
+@router.command(model="EUYieldCurve")
+def eu_yield_curve(
+    cc: CommandContext,
+    provider_choices: ProviderChoices,
+    standard_params: StandardParams,
+    extra_params: ExtraParams,
+) -> OBBject:
+    """Euro Area Yield Curve.
+
+    Gets euro area yield curve data from ECB.
+
+    The graphic depiction of the relationship between the yield on bonds of the same credit quality but different
+    maturities is known as the yield curve. In the past, most market participants have constructed yield curves from
+    the observations of prices and yields in the Treasury market. Two reasons account for this tendency. First,
+    Treasury securities are viewed as free of default risk, and differences in creditworthiness do not affect yield
+    estimates. Second, as the most active bond market, the Treasury market offers the fewest problems of illiquidity
+    or infrequent trading. The key function of the Treasury yield curve is to serve as a benchmark for pricing bonds
+    and setting yields in other sectors of the debt market.
+
+    It is clear that the market’s expectations of future rate changes are one important determinant of the
+    yield-curve shape. For example, a steeply upward-sloping curve may indicate market expectations of near-term Fed
+    tightening or of rising inflation. However, it may be too restrictive to assume that the yield differences across
+    bonds with different maturities only reflect the market’s rate expectations. The well-known pure expectations
+    hypothesis has such an extreme implication. The pure expectations hypothesis asserts that all government bonds
+    have the same near-term expected return (as the nominally riskless short-term bond) because the return-seeking
+    activity of risk-neutral traders removes all expected return differentials across bonds.
+    """
+    return OBBject(results=Query(**locals()).execute())
+
+
+@router.command(model="TreasuryRates")
+def treasury(
+    cc: CommandContext,
+    provider_choices: ProviderChoices,
+    standard_params: StandardParams,
+    extra_params: ExtraParams,
+) -> OBBject[BaseModel]:
+    """Treasury Rates. Treasury rates data."""
+    return OBBject(results=Query(**locals()).execute())
