+++ openbb_platform/extensions/fixedincome/openbb_fixedincome/spreads/__init__.py
@@ -1,0 +1,1 @@
+"""Initialize the Fixed Income Spreads module."""
