+++ openbb_platform/openbb/package/module_map.json
@@ -56,27 +56,30 @@
     "etf_search": "/etf/search",
     "etf_sectors": "/etf/sectors",
     "fixedincome": "/fixedincome",
-    "fixedincome_cp": "/fixedincome/cp",
-    "fixedincome_dwpcr": "/fixedincome/dwpcr",
-    "fixedincome_ecb_interest_rates": "/fixedincome/ecb_interest_rates",
-    "fixedincome_eu_yield_curve": "/fixedincome/eu_yield_curve",
-    "fixedincome_ffrmc": "/fixedincome/ffrmc",
-    "fixedincome_hqm": "/fixedincome/hqm",
-    "fixedincome_ice_bofa": "/fixedincome/ice_bofa",
-    "fixedincome_moody": "/fixedincome/moody",
+    "fixedincome_corporate": "/fixedincome/corporate",
+    "fixedincome_corporate_commercial_paper": "/fixedincome/corporate/commercial_paper",
+    "fixedincome_corporate_hqm": "/fixedincome/corporate/hqm",
+    "fixedincome_corporate_ice_bofa": "/fixedincome/corporate/ice_bofa",
+    "fixedincome_corporate_moody": "/fixedincome/corporate/moody",
+    "fixedincome_corporate_spot_rates": "/fixedincome/corporate/spot_rates",
+    "fixedincome_government": "/fixedincome/government",
+    "fixedincome_government_eu_yield_curve": "/fixedincome/government/eu_yield_curve",
+    "fixedincome_government_treasury": "/fixedincome/government/treasury",
+    "fixedincome_government_us_yield_curve": "/fixedincome/government/us_yield_curve",
     "fixedincome_rate": "/fixedincome/rate",
     "fixedincome_rate_ameribor": "/fixedincome/rate/ameribor",
+    "fixedincome_rate_dpcredit": "/fixedincome/rate/dpcredit",
+    "fixedincome_rate_ecb": "/fixedincome/rate/ecb",
     "fixedincome_rate_effr": "/fixedincome/rate/effr",
     "fixedincome_rate_effr_forecast": "/fixedincome/rate/effr_forecast",
     "fixedincome_rate_estr": "/fixedincome/rate/estr",
     "fixedincome_rate_iorb": "/fixedincome/rate/iorb",
     "fixedincome_rate_sonia": "/fixedincome/rate/sonia",
     "fixedincome_sofr": "/fixedincome/sofr",
-    "fixedincome_spot": "/fixedincome/spot",
-    "fixedincome_tbffr": "/fixedincome/tbffr",
-    "fixedincome_tmc": "/fixedincome/tmc",
-    "fixedincome_treasury": "/fixedincome/treasury",
-    "fixedincome_yield_curve": "/fixedincome/yield_curve",
+    "fixedincome_spreads": "/fixedincome/spreads",
+    "fixedincome_spreads_tmc": "/fixedincome/spreads/tmc",
+    "fixedincome_spreads_tmc_effr": "/fixedincome/spreads/tmc_effr",
+    "fixedincome_spreads_treasury_effr": "/fixedincome/spreads/treasury_effr",
     "futures": "/futures",
     "futures_curve": "/futures/curve",
     "futures_load": "/futures/load",
